Add peer retrieval code (iteration) + organize accross files
